Refactor user sign up  form validation
Create server.validation.go file to contain all validations
